add docker config files and refactor for deploying
