+++ app/app_models/detection_result.py
