+++ app/app_utils/model_manager.py
@@ -1,6 +1,8 @@
 import os
 from loguru import logger
-from app.models.yolo_model import YOLOModel
+from app_models.yolo_model import YOLOModel
+
+logger.add("logs/model_manager.log", rotation="5 MB", level="ERROR")
 
 
 class ModelManager:
@@ -8,11 +10,15 @@ class ModelManager:
 
     def __init__(self, model_path: str):
         """Initialize the model manager with a model path."""
+        if not os.path.exists(model_path):
+            logger.error(f"Model path does not exist: {model_path}")
+            raise ValueError("Model path does not exist")
         self.model = YOLOModel(model_path)
 
     def set_model(self, model_path: str):
         """Set a new model for detection."""
         if not os.path.exists(model_path):
+            logger.error(f"Model path does not exist: {model_path}")
             raise ValueError("Model path does not exist")
         self.model = YOLOModel(model_path)
         logger.info(f"Model set to {model_path}")
