+++ frontend/public/training_progress.csv
