+++ pycodestyle.py
@@ -1,6 +1,10 @@
+import os
 import sys
 import subprocess
-import os
+
+# Добавьте пути к app и frontend
+sys.path.insert(0, os.path.join(os.getcwd(), 'app'))
+sys.path.insert(0, os.path.join(os.getcwd(), 'frontend'))
 
 
 def run_flake8(script_path):
@@ -14,9 +18,9 @@ def run_flake8(script_path):
 
 def run_pylint(script_path):
     """Run pylint on the specified script."""
-    os.environ['PYTHONPATH'] = os.getcwd()
+    pythonpath = os.pathsep.join([os.path.join(os.getcwd(), 'app'), os.path.join(os.getcwd(), 'frontend')])
     pylint_command = (
-        f"pylint {script_path} "
+        f"PYTHONPATH={pythonpath} pylint {script_path} "
         "--max-line-length=120 "
         "--disable=C0103,C0114,C0115 "
         "--generated-members=cv2.*"
